# packages/lib/src/docs/openapi.yaml
openapi: 3.1.0
info:
  title: '@oas-tools/oas-telemetry'
  version: 0.7.0-alpha.2
  license:
    name: Apache-2.0
  description: |-
    Automatically collects and stores telemetry data in memory using OpenTelemetry.
    Just import the Express middleware to access it through the built-in REST API or web UI—no extra setup required.
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: oas-tlm-access-token
security:
  - CookieAuth: []
servers:
  - url: "{protocol}://{host}:{port}/{basePath}"
    variables:
      protocol:
        default: http
        enum:
          - http
          - https
      host:
        default: localhost
      port:
        default: "3000"
      basePath:
        default: telemetry
        description: Base path for all endpoints. Can be customized, e.g. "myCustomTelemetryPath"
...
paths:
...
  /utils/generate-log:
    get:
      summary: Generate a sample log
      tags:
        - utils
      parameters:
        - name: log
          in: query
          required: false
          schema:
            type: string
          description: The log message to generate. Defaults to "Default log message" if not specified.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Log generated
                  log:
                    type: string
                    example: Hello World!
  /utils/wait/{seconds}:
    get:
      summary: Wait for the specified seconds
      tags:
        - utils
      parameters:
        - name: seconds
          in: path
          required: true
          schema:
            type: number
          description: Number of seconds to wait. Defaults to 1 if not specified.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  waited:
                    type: number
                    example: 2
...